requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.1
nepse @ git+https://github.com/basic-bgnr/NepseUnofficialApi@901d3c8c2f5ab9e80c19a0d42ba2ab0a4871f35f
numpy==2.1.2
PyJWT==2.9.0
pywasm==1.0.8
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
Werkzeug==3.0.4
